added redux state management for group requests
